# Conflicts:
#	backend-nest/src/question/service/question.service.ts
#	backend-nest/src/questionnary/controller/questionnary.controller.ts
#	backend-nest/src/questionnary/dto/questionnaryCreate.dto.ts
#	backend-nest/src/questionnary/entity/questionnary.entity.ts
#	backend-nest/src/questionnary/service/questionnary.service.ts
#	frontend-vue/src/api/question.js
#	frontend-vue/src/api/questionnary.js
#	frontend-vue/src/stores/questionnaryStore.js

+++ backend-nest/src/questionnary/controller/questionnary.controller.ts
@@ -14,6 +14,13 @@ import { QuestionnaryCreateDto } from '../dto/questionnaryCreate.dto';
 import { QuestionCreateDto } from '../../question/dto/questionCreate.dto';
 import { QuestionService } from '../../question/service/question.service';
 import { TagDto } from '../../question/dto/tag.dto';
+import { Req } from '@nestjs/common/decorators/http/route-params.decorator';
+import { UserRequest } from '../../auth/config/user.request';
+import { UserType } from '../../user/constants/userType.constant';
+import { Roles } from '../../decorators/roles.decorator';
+import { Teacher } from '../../user/entity/teacher.entity';
+import { NotAuthorException } from '../exception/notAuthor.exception';
+import { QuestionnaryNotFoundException } from '../exception/questionnaryNotFound.exception';
 
 @Controller('questionnary')
 export class QuestionnaryController {
@@ -49,42 +56,78 @@ export class QuestionnaryController {
     return this.questionService.getTags(id);
   }
 
+  @Roles([UserType.TEACHER])
   @Post('/create')
   createQuestionnary(
+    @Req() request: UserRequest,
     @Body(new ValidationPipe()) questionnaryDto: QuestionnaryCreateDto,
   ) {
-    return this.questionnaryService.createQuestionnary(questionnaryDto);
+    return this.questionnaryService.createQuestionnary(
+      questionnaryDto,
+      <Teacher>request.user,
+    );
   }
 
+  @Roles([UserType.TEACHER])
   @Delete('/:id')
-  async deleteQuestionnary(@Param('id', ParseIntPipe) idQuestionnary: number) {
+  async deleteQuestionnary(
+    @Req() request: UserRequest,
+    @Param('id', ParseIntPipe) idQuestionnary: number,
+  ) {
+    if (!(await this.questionnaryService.questionnaryExists(idQuestionnary)))
+      throw new QuestionnaryNotFoundException();
+    if (
+      !(await this.questionnaryService.isQuestionnaryFromTeacher(
+        idQuestionnary,
+        request.user as Teacher,
+      ))
+    )
+      throw new NotAuthorException();
     return this.questionnaryService.deleteQuestionnary(idQuestionnary);
   }
 
   @Get('/:id')
-  selectQuestionnary(@Param('id', ParseIntPipe) idQuestionnary: number) {
+  async selectQuestionnary(@Param('id', ParseIntPipe) idQuestionnary: number) {
+    if (!(await this.questionnaryService.questionnaryExists(idQuestionnary)))
+      throw new QuestionnaryNotFoundException();
     return this.questionnaryService.findQuestionnary(idQuestionnary);
   }
 
-  @Get('/user/:idUser')
-  selectQuestionnariesFromUser(@Param('idUser', ParseIntPipe) idUser: number) {
-    return this.questionnaryService.findQuestionnariesFromIdUser(idUser);
+  @Roles([UserType.TEACHER])
+  @Get()
+  selectQuestionnaryList(@Req() request: UserRequest) {
+    return this.questionnaryService.findQuestionnariesFromIdUser(
+      request.user as Teacher,
+    );
   }
 
   @Get('/:id/question/')
-  selectQuestionsFromQuestionnary(
+  async selectQuestionsFromQuestionnary(
     @Param('id', ParseIntPipe) idQuestionnary: number,
   ) {
+    if (!(await this.questionnaryService.questionnaryExists(idQuestionnary)))
+      throw new QuestionnaryNotFoundException();
     return this.questionnaryService.findQuestionsFromIdQuestionnary(
       idQuestionnary,
     );
   }
 
+  @Roles([UserType.TEACHER])
   @Post('/:id/question/')
   async addQuestion(
     @Param('id', ParseIntPipe) idQuestionnary: number,
     @Body(new ValidationPipe()) questionDto: QuestionCreateDto,
+    @Req() request: UserRequest,
   ) {
+    if (!(await this.questionnaryService.questionnaryExists(idQuestionnary)))
+      throw new QuestionnaryNotFoundException();
+    if (
+      !(await this.questionnaryService.isQuestionnaryFromTeacher(
+        idQuestionnary,
+        request.user as Teacher,
+      ))
+    )
+      throw new NotAuthorException();
     return this.questionnaryService.addQuestion(idQuestionnary, questionDto);
   }
 
@@ -93,20 +136,42 @@ export class QuestionnaryController {
     return await this.questionService.getQuestionPrivateBank(idUser);
   }
 
+  @Roles([UserType.TEACHER])
   @Delete('/:id/question/:idQuestion')
   async deleteQuestion(
     @Param('id', ParseIntPipe) idQuestionnary: number,
     @Param('idQuestion', ParseIntPipe) idQuestion: number,
+    @Req() request: UserRequest,
   ) {
+    if (!(await this.questionnaryService.questionnaryExists(idQuestionnary)))
+      throw new QuestionnaryNotFoundException();
+    if (
+      !(await this.questionnaryService.isQuestionnaryFromTeacher(
+        idQuestionnary,
+        request.user as Teacher,
+      ))
+    )
+      throw new NotAuthorException();
     return this.questionnaryService.deleteQuestion(idQuestionnary, idQuestion);
   }
 
+  @Roles([UserType.TEACHER])
   @Patch('/:id/question/:idQuestion')
   async modifyQuestion(
     @Param('id', ParseIntPipe) idQuestionnary: number,
     @Param('idQuestion', ParseIntPipe) idQuestion: number,
     @Body(new ValidationPipe()) questionDto: QuestionCreateDto,
+    @Req() request: UserRequest,
   ) {
+    if (!(await this.questionnaryService.questionnaryExists(idQuestionnary)))
+      throw new QuestionnaryNotFoundException();
+    if (
+      !(await this.questionnaryService.isQuestionnaryFromTeacher(
+        idQuestionnary,
+        request.user as Teacher,
+      ))
+    )
+      throw new NotAuthorException();
     return this.questionnaryService.modifyQuestion(
       idQuestionnary,
       idQuestion,
@@ -114,15 +179,26 @@ export class QuestionnaryController {
     );
   }
 
+  @Roles([UserType.TEACHER])
   @Patch('/:id')
-  modifyQuestionnary(
+  async modifyQuestionnary(
     @Param('id', ParseIntPipe) idQuestionnary: number,
     @Body() body: { questionnaryName: string },
+    @Req() request: UserRequest,
   ) {
+    if (!(await this.questionnaryService.questionnaryExists(idQuestionnary)))
+      throw new QuestionnaryNotFoundException();
+    if (
+      !(await this.questionnaryService.isQuestionnaryFromTeacher(
+        idQuestionnary,
+        request.user as Teacher,
+      ))
+    )
+      throw new NotAuthorException();
     return this.questionnaryService.modifyQuestionnary(
       idQuestionnary,
       body.questionnaryName,
-      //body.author,
+      request.user as Teacher,
     );
   }
 }
